requirements.txt
streamlit
google-cloud-bigquery
google-cloud-aiplatform
streamlit-calendar
pytest
plotly
vertexai
streamlit-cookies-manager

